storygrid move ui
